chore: move combineLatest into separate file
