+++ src/operators/index.ts
@@ -9,7 +9,8 @@ export { bufferWhen } from '../internal/operators/bufferWhen';
 export { catchError } from '../internal/operators/catchError';
 export { combineAll } from '../internal/operators/combineAll';
 export { combineLatestAll } from '../internal/operators/combineLatestAll';
-export { combineLatest, combineLatestWith } from '../internal/operators/combineLatestWith';
+export { combineLatest } from '../internal/operators/combineLatest';
+export { combineLatestWith } from '../internal/operators/combineLatestWith';
 export { concatAll } from '../internal/operators/concatAll';
 export { concatMap } from '../internal/operators/concatMap';
 export { concatMapTo } from '../internal/operators/concatMapTo';
